+++ Allure.Net.Commons.Tests/ConcurrencyTests.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Xml.Linq;
 using NUnit.Framework;
 using NUnit.Framework.Internal;
