+++ Allure.Net.Commons/Storage/AllureStorage.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Threading;
+﻿using System.Collections.Concurrent;
 
 #nullable enable
 
@@ -11,269 +6,22 @@ namespace Allure.Net.Commons.Storage
 {
     internal class AllureStorage
     {
-        readonly AsyncLocal<AllureContext> context = new();
-
-        public AllureContext CurrentContext
-        {
-            get => this.context.Value ??= new();
-            set => this.context.Value = value
-                ?? throw new ArgumentNullException(
-                    nameof(this.CurrentContext)
-                );
-        }
-
-        public AllureStorage()
-        {
-            this.CurrentContext = new();
-        }
-
-        public TestResultContainer? CurrentTestContainerOrNull =>
-            this.CurrentContext.ContainerContext.FirstOrDefault();
-
-        public TestResultContainer CurrentTestContainer =>
-            this.CurrentContext.CurrentContainer;
-
-        public FixtureResult CurrentFixture =>
-            this.CurrentContext.CurrentFixture;
-
-        public TestResult CurrentTest =>
-            this.CurrentContext.CurrentTest;
-
-        public ExecutableItem CurrentStepContainer =>
-            this.CurrentContext.CurrentStepContainer;
-
-        public StepResult CurrentStep =>
-            this.CurrentContext.CurrentStep;
-
-        public void PutTestContainer(TestResultContainer container) =>
-            this.UpdateContext(
-                c => c.WithContainer(container)
-            );
-
-        public void RemoveTestContainer() => this.UpdateContext(
-            c => c.WithNoLastContainer()
-        );
-
-        public void PutFixture(FixtureResult fixture) =>
-            this.UpdateContext(c => c.WithFixtureContext(fixture));
-
-        public void RemoveFixture() =>
-            this.UpdateContext(c => c.WithNoFixtureContext());
-
-        public void PutTestCase(TestResult testResult) =>
-            this.UpdateContext(
-                c => c.WithTestContext(testResult)
-            );
-
-        public void RemoveTestCase() =>
-            this.UpdateContext(
-                c => c.WithNoTestContext()
-            );
-
-        public void PutStep(StepResult stepResult) =>
-            this.UpdateContext(
-                c => c.WithStep(stepResult)
-            );
-
-        public void RemoveStep() => this.UpdateContext(
-            c => c.WithNoLastStep()
-        );
-
-        void UpdateContext(Func<AllureContext, AllureContext> updateFn)
-        {
-            this.CurrentContext = updateFn(this.CurrentContext);
-        }
-
-        #region Obsoleted
-
-        [Obsolete]
         readonly ConcurrentDictionary<string, object> storage = new();
 
-        [Obsolete]
         public T Get<T>(string uuid)
         {
             return (T)storage[uuid];
         }
 
-        [Obsolete]
         public T Put<T>(string uuid, T item) where T : notnull
         {
             return (T)storage.GetOrAdd(uuid, item);
         }
 
-        [Obsolete]
         public T Remove<T>(string uuid)
         {
             storage.TryRemove(uuid, out var value);
             return (T)value;
         }
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void PutTestContainer(
-            string uuid,
-            TestResultContainer container
-        ) => this.PutAndUpdateContext(
-            uuid,
-            container,
-            c => c.WithContainer(container)
-        );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void RemoveTestContainer(string uuid) =>
-            this.RemoveAndUpdateContext<TestResultContainer>(
-                uuid,
-                c => ContextWithNoContainer(c, uuid)
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void PutFixture(string uuid, FixtureResult fixture) =>
-            this.PutAndUpdateContext(
-                uuid,
-                fixture,
-                c => c.WithFixtureContext(fixture)
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void RemoveFixture(string uuid)
-            => this.RemoveAndUpdateContext<FixtureResult>(
-                uuid,
-                c => c.WithNoFixtureContext()
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void PutTestCase(string uuid, TestResult testResult) =>
-            this.PutAndUpdateContext(
-                uuid,
-                testResult,
-                c => c.WithTestContext(testResult)
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void RemoveTestCase(string uuid) =>
-            this.RemoveAndUpdateContext<TestResult>(
-                uuid,
-                c => c.WithNoTestContext()
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void PutStep(string uuid, StepResult stepResult) =>
-            this.PutAndUpdateContext(
-                uuid,
-                stepResult,
-                c => c.WithStep(stepResult)
-            );
-
-        [Obsolete(
-            "Storage methods with explicit uuids are obsolete. Use " +
-                "their counterparts without uuids to manipulate the context."
-        )]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public void RemoveStep(string uuid) =>
-            this.RemoveAndUpdateContext<StepResult>(
-                uuid,
-                c => this.ContextWithNoStep(c, uuid)
-            );
-
-        [Obsolete]
-        void PutAndUpdateContext<T>(
-            string uuid,
-            T value,
-            Func<AllureContext, AllureContext> updateFn
-        ) where T : notnull
-        {
-            this.Put(uuid, value);
-            this.UpdateContext(updateFn);
-        }
-
-        [Obsolete]
-        void RemoveAndUpdateContext<T>(string uuid, Func<AllureContext, AllureContext> updateFn)
-        {
-            this.UpdateContext(updateFn);
-            this.Remove<T>(uuid);
-        }
-
-        [Obsolete]
-        static AllureContext ContextWithNoContainer(
-            AllureContext context,
-            string uuid
-        )
-        {
-            var containersToPushAgain = new Stack<TestResultContainer>();
-            while (context.CurrentContainer.uuid != uuid)
-            {
-                containersToPushAgain.Push(context.CurrentContainer);
-                context = context.WithNoLastContainer();
-                if (context.ContainerContext.IsEmpty)
-                {
-                    throw new InvalidOperationException(
-                        $"Container {uuid} is not in the current context"
-                    );
-                }
-            }
-            context = context.WithNoLastContainer();
-            while (containersToPushAgain.Any())
-            {
-                context = context.WithContainer(
-                    containersToPushAgain.Pop()
-                );
-            }
-            return context;
-        }
-
-        [Obsolete]
-        AllureContext ContextWithNoStep(AllureContext context, string uuid)
-        {
-            var stepResult = this.Get<StepResult>(uuid);
-            var stepsToPushAgain = new Stack<StepResult>();
-            while (!ReferenceEquals(context.CurrentStep, stepResult))
-            {
-                stepsToPushAgain.Push(context.CurrentStep);
-                context = context.WithNoLastStep();
-                if (context.StepContext.IsEmpty)
-                {
-                    throw new InvalidOperationException(
-                        $"Step {stepResult.name} is not in the current context"
-                    );
-                }
-            }
-            context = context.WithNoLastStep();
-            while (stepsToPushAgain.Any())
-            {
-                context = context.WithStep(
-                    stepsToPushAgain.Pop()
-                );
-            }
-            return context;
-        }
-
-        #endregion
     }
 }
