+++ Allure.SpecFlowPlugin/PluginHelper.cs
@@ -6,7 +6,6 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using Allure.Net.Commons;
-using Allure.Net.Commons.Storage;
 using Newtonsoft.Json.Linq;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Bindings;
